+++ src/hooks/useEtherspotBalances.ts
@@ -1,5 +1,6 @@
 import { useEtherspot } from '@etherspot/react-etherspot';
 import { AccountTypes, AccountBalance } from 'etherspot';
+import { useMemo } from 'react';
 
 interface IEtherspotBalancesHook {
   getAccountBalances: (accountAddress?: string) => Promise<AccountBalance[]>;
@@ -7,27 +8,32 @@ interface IEtherspotBalancesHook {
 
 /**
  * Hook to fetch account balances
- * @param chainId {number} - Chain ID
+ * @param chainId {number | undefined} - Chain ID
  * @returns {IEtherspotBalancesHook} - hook method to fetch Etherspot account balances
  */
-const useEtherspotBalances = (chainId: number = 1): IEtherspotBalancesHook => {
-  const { connect, getSdkForChainId } = useEtherspot();
+const useEtherspotBalances = (chainId?: number): IEtherspotBalancesHook => {
+  const { connect, getSdkForChainId, chainId: defaultChainId } = useEtherspot();
+
+  const historyChainId = useMemo(() => {
+    if (chainId) return chainId;
+    return defaultChainId;
+  }, [chainId, defaultChainId]);
 
   const getAccountBalances = async (accountAddress?: string) => {
-    const sdkForChainId = getSdkForChainId(chainId);
+    const sdkForChainId = getSdkForChainId(historyChainId);
     if (!sdkForChainId) {
-      console.warn(`Unable to get SDK for chain ID ${chainId}`);
+      console.warn(`Unable to get SDK for chain ID ${historyChainId}`);
       return [];
     }
 
     if (sdkForChainId?.state?.account?.type !== AccountTypes.Contract) {
-      await connect(chainId);
+      await connect(historyChainId);
     }
 
     try {
       const { items } = await sdkForChainId.getAccountBalances({
         account: accountAddress ?? sdkForChainId.state.account.address,
-        chainId,
+        chainId: historyChainId,
       });
 
       return items;