+++ src/hooks/useEtherspotHistory.ts
@@ -1,5 +1,6 @@
 import { AccountTypes, Transaction } from 'etherspot';
 import { useEtherspot } from '@etherspot/react-etherspot';
+import { useMemo } from 'react';
 
 interface IEtherspotHistoryHook {
   getAccountTransactions: (accountAddress?: string) => Promise<Transaction[]>;
@@ -8,18 +9,23 @@ interface IEtherspotHistoryHook {
 
 /**
  * Hook to fetch Etherspot account transactions history
- * @param chainId {number} - Chain ID
+ * @param chainId {number | undefined} - Chain ID
  * @returns {IEtherspotHistoryHook} - hook methods to fetch Etherspot transactions history
  */
-const useEtherspotHistory = (chainId: number = 1): IEtherspotHistoryHook => {
-  const { connect, getSdkForChainId } = useEtherspot();
+const useEtherspotHistory = (chainId: number): IEtherspotHistoryHook => {
+  const { connect, getSdkForChainId, chainId: defaultChainId } = useEtherspot();
+
+  const historyChainId = useMemo(() => {
+    if (chainId) return chainId;
+    return defaultChainId;
+  }, [chainId, defaultChainId]);
 
   const getAccountTransactions = async (accountAddress?: string): Promise<Transaction[]> => {
-    const sdkForChainId = getSdkForChainId(chainId);
+    const sdkForChainId = getSdkForChainId(historyChainId);
     if (!sdkForChainId) return [];
 
     if (sdkForChainId?.state?.account?.type !== AccountTypes.Contract) {
-      await connect(chainId);
+      await connect(historyChainId);
     }
 
     let transactions: Transaction[] = [];
@@ -41,11 +47,11 @@ const useEtherspotHistory = (chainId: number = 1): IEtherspotHistoryHook => {
   };
 
   const getAccountTransaction = async (hash: string): Promise<Transaction | undefined> => {
-    const sdkForChainId = getSdkForChainId(chainId);
+    const sdkForChainId = getSdkForChainId(historyChainId);
     if (!sdkForChainId) return;
 
     if (sdkForChainId?.state?.account?.type !== AccountTypes.Contract) {
-      await connect(chainId);
+      await connect(historyChainId);
     }
 
     try {